app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
kafka-python==2.0.2
python-dotenv==1.0.0
minio==7.1.17
pyarrow==14.0.1
pandas==2.1.3
pyspark==3.5.0
clickhouse-driver==0.2.6
six==1.16.0
requests==2.31.0
apache-airflow==2.8.1
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-apache-kafka==1.3.1